requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
pydantic==1.8.1
python-dotenv==0.17.1
starlette==0.13.6
typing-extensions==3.10.0.0
uvicorn==0.13.4
boto3==1.16.52
python-magic
redis
requests